Refactor homepage.js, remove apartment.js
