--- src/components/home/homepage.js
@@ -1,39 +1,0 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
-import { useSelector, useDispatch } from 'react-redux';
-import { singleApartments } from '../../redux/apartment/apartment';
-
-const Home = () => {
-  const dispatch = useDispatch();
-  const apartmentss = useSelector((state) => state.apartment);
-  const { apartments } = apartmentss;
-  if (!apartments[0]) {
-    return (
-      <h1>Loading</h1>
-    );
-  }
-  return (
-    <section>
-      <h1 className="apartment-heading text-center mt-3 text-uppercase">Latest Home</h1>
-      <div className="row">
-        {
-      apartments[0].map((item) => (
-        <div key={item.id} className="col-12 col-md-6 col-lg-4">
-          <Link style={{ textDecoration: 'none' }} to={`${item.id}`} className="m-3" onClick={() => dispatch(singleApartments(item.id))}>
-            <div className="card mx-3">
-              <img className="apartment-image hover_effect center-block" src={item.image} alt={item.name} />
-              <h4 className="my-2 text-center apartmentname">{item.name}</h4>
-              <p className="description font-weight-light text-center ms-3">
-                {item.description}
-              </p>
-            </div>
-          </Link>
-        </div>
-      ))
-    }
-      </div>
-    </section>
-  );
-};
-
-export default Home;
